review comments to move to jakarta.data.metamodel.restrict
