+++ api/src/main/java/jakarta/data/metamodel/TextAttribute.java
@@ -18,8 +18,8 @@
 package jakarta.data.metamodel;
 
 import jakarta.data.Sort;
-import jakarta.data.restrict.Restrict;
-import jakarta.data.restrict.TextRestriction;
+import jakarta.data.metamodel.restrict.Restrict;
+import jakarta.data.metamodel.restrict.TextRestriction;
 
 /**
  * Represents an textual entity attribute in the {@link StaticMetamodel}.
