+++ api/src/main/java/jakarta/data/metamodel/restrict/BasicRestriction.java
@@ -15,20 +15,15 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.restrict;
+package jakarta.data.metamodel.restrict;
 
-public interface TextRestriction<T> extends BasicRestriction<T> {
-    TextRestriction<T> ignoreCase();
+public interface BasicRestriction<T> extends Restriction<T> {
+    String attribute();
 
-    // TODO can mention in the JavaDoc that a value of true will be ignored
-    // if the database is not not capable of case sensitive comparisons
-    boolean isCaseSensitive();
-
-    boolean isEscaped();
+    Operator comparison();
 
     @Override
-    TextRestriction<T> negate();
+    BasicRestriction<T> negate();
 
-    @Override
-    String value();
+    Object value();
 }
