+++ api/src/main/java/jakarta/data/metamodel/restrict/BasicRestrictionRecord.java
@@ -15,7 +15,7 @@
  *
  *  SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.restrict;
+package jakarta.data.metamodel.restrict;
 
 // Internal implementation class.
 // The proper way for users to obtain instances is via
@@ -23,38 +23,20 @@ package jakarta.data.restrict;
 
 import java.util.Objects;
 
-record TextRestrictionRecord<T>(
+record BasicRestrictionRecord<T>(
         String attribute,
         Operator comparison,
-        boolean isCaseSensitive,
-        boolean isEscaped,
-        String value) implements TextRestriction<T> {
+        Object value) implements BasicRestriction<T> {
 
-    TextRestrictionRecord {
+    BasicRestrictionRecord {
         Objects.requireNonNull(attribute, "Attribute must not be null");
     }
 
-    TextRestrictionRecord(String attributeName, Operator comparison, boolean escaped, String value) {
-        this(attributeName, comparison, true, escaped, value);
-    }
-
-    TextRestrictionRecord(String attributeName, Operator comparison, String value) {
-        this(attributeName, comparison, true, false, value);
-    }
-
     @Override
-    public TextRestriction<T> ignoreCase() {
-        return new TextRestrictionRecord<>(attribute, comparison, false, isEscaped, value);
-    }
-
-    @Override
-    public TextRestriction<T> negate() {
-
-        return new TextRestrictionRecord<>(
+    public BasicRestriction<T> negate() {
+        return new BasicRestrictionRecord<>(
                 attribute,
                 comparison.negate(),
-                isCaseSensitive,
-                isEscaped,
                 value);
     }
 }
