+++ api/src/main/java/jakarta/data/metamodel/restrict/CompositeRestrictionRecord.java
@@ -15,7 +15,7 @@
  *
  *  SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.restrict;
+package jakarta.data.metamodel.restrict;
 
 import java.util.List;
