+++ api/src/main/java/jakarta/data/metamodel/restrict/Restriction.java
@@ -15,22 +15,8 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.restrict;
+package jakarta.data.metamodel.restrict;
 
-import java.util.List;
-
-public interface CompositeRestriction<T> extends Restriction<T> {
-    boolean isNegated();
-
-    @Override
-    CompositeRestriction<T> negate();
-
-    List<Restriction<T>> restrictions();
-
-    Type type();
-
-    enum Type {
-        ALL,
-        ANY
-    }
+public interface Restriction<T> {
+    Restriction<T> negate();
 }
