+++ api/src/main/java/jakarta/data/metamodel/restrict/TextRestriction.java
@@ -15,15 +15,20 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.restrict;
+package jakarta.data.metamodel.restrict;
 
-public interface BasicRestriction<T> extends Restriction<T> {
-    String attribute();
+public interface TextRestriction<T> extends BasicRestriction<T> {
+    TextRestriction<T> ignoreCase();
 
-    Operator comparison();
+    // TODO can mention in the JavaDoc that a value of true will be ignored
+    // if the database is not not capable of case sensitive comparisons
+    boolean isCaseSensitive();
+
+    boolean isEscaped();
 
     @Override
-    BasicRestriction<T> negate();
+    TextRestriction<T> negate();
 
-    Object value();
+    @Override
+    String value();
 }
