+++ api/src/main/java/jakarta/data/metamodel/restrict/TextRestrictionRecord.java
@@ -15,7 +15,7 @@
  *
  *  SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.restrict;
+package jakarta.data.metamodel.restrict;
 
 // Internal implementation class.
 // The proper way for users to obtain instances is via
@@ -23,20 +23,38 @@ package jakarta.data.restrict;
 
 import java.util.Objects;
 
-record BasicRestrictionRecord<T>(
+record TextRestrictionRecord<T>(
         String attribute,
         Operator comparison,
-        Object value) implements BasicRestriction<T> {
+        boolean isCaseSensitive,
+        boolean isEscaped,
+        String value) implements TextRestriction<T> {
 
-    BasicRestrictionRecord {
+    TextRestrictionRecord {
         Objects.requireNonNull(attribute, "Attribute must not be null");
     }
 
+    TextRestrictionRecord(String attributeName, Operator comparison, boolean escaped, String value) {
+        this(attributeName, comparison, true, escaped, value);
+    }
+
+    TextRestrictionRecord(String attributeName, Operator comparison, String value) {
+        this(attributeName, comparison, true, false, value);
+    }
+
     @Override
-    public BasicRestriction<T> negate() {
-        return new BasicRestrictionRecord<>(
+    public TextRestriction<T> ignoreCase() {
+        return new TextRestrictionRecord<>(attribute, comparison, false, isEscaped, value);
+    }
+
+    @Override
+    public TextRestriction<T> negate() {
+
+        return new TextRestrictionRecord<>(
                 attribute,
                 comparison.negate(),
+                isCaseSensitive,
+                isEscaped,
                 value);
     }
 }
