+++ api/src/main/java/module-info.java
@@ -1069,10 +1069,10 @@ module jakarta.data {
     exports jakarta.data;
     exports jakarta.data.metamodel;
     exports jakarta.data.metamodel.impl;
+    exports jakarta.data.metamodel.restrict;
     exports jakarta.data.page;
     exports jakarta.data.page.impl;
     exports jakarta.data.repository;
-    exports jakarta.data.restrict;
     exports jakarta.data.exceptions;
     opens jakarta.data.repository;
     exports jakarta.data.spi;
